Added permissions

+++ commands/Moderation/purge.js
@@ -1,4 +1,5 @@
 const { SlashCommandBuilder } = require("@discordjs/builders");
+const { PermissionFlagsBits } = require("discord.js");
 
 module.exports = {
   data: new SlashCommandBuilder()
@@ -12,6 +13,15 @@ module.exports = {
     ),
   async execute(interaction) {
     const amount = interaction.options.getInteger("amount");
+
+    if (
+      !interaction.member.permissions.has(PermissionFlagsBits.ManageMessages)
+    ) {
+      return interaction.reply({
+        content: "You do not have the `MANAGE_MESSAGES` permission!",
+        ephemeral: true,
+      });
+    }
 
     interaction.channel.bulkDelete(amount).then((messages) =>
       interaction.reply({
